refactor(errs): split Int(int)+Int64(int64) per slog/zap convention
Fixes the Field API ergonomics flagged in the Go-idioms audit
(.claude/contexts/go-idioms-errs-field-api.md). Every errs.Int call
site was writing int64(...) because the signature was
Int(key string, val int64) — odd one out vs the ecosystem:

  log/slog:   func Int(key string, value int) Attr
  zap:        func Int(key string, val int) Field
  errs (us):  func Int(key string, val int64)  ← forced cast everywhere

Changes:

- errs.Int now takes int and widens internally (matches slog + zap).
- errs.Int64 added for callers who already hold an int64; no data loss.
- 14 production call sites drop their int64(...) wrapper:
    errs.Int("bytes", int64(len(p)))        → errs.Int("bytes", len(p))
    errs.Int("level", int64(rec.Level))     → errs.Int("level", int(rec.Level))
    errs.Int("cap", int64(maxYAMLBytes))    → errs.Int("cap", maxYAMLBytes)
  Level-to-int narrowing is still one cast (Level is int8) but it is
  the idiomatic Go spelling.

Performance: NO regression. int → int64 widening is a 1-cycle MOVSX
on x86-64 / ARM64; on a 64-bit platform int and int64 have identical
machine representation. The gain is ergonomic (cast noise gone from
14 call sites) — the 'tue les performance' concern from the review
is defused by the compiler.

Also in this commit, three smaller cleanups from the same audit:

1. durationBits(time.Duration) uint64 helper in core/logger/value.go
   isolates the mandatory two-step conversion
   uint64(int64(time.Duration)) into a named function so the call
   site stops carrying the chained cast.

2. internal/core/codec/registry.go: add a godoc block explaining why
   sync.Map is deliberate over RWMutex+map (append-once-read-many
   matches the stdlib sync.Map sweet spot; reverting would regress
   hot-path Lookup latency).

3. internal/core/codec/codes.go constants retyped as errs.Code to
   close the last KTN-CONST-TYPED drift on this file.

Audit context: .claude/contexts/go-idioms-errs-field-api.md

Tests: bazel test --config=race //... → 32/32 pass.

+++ internal/service/codec/msgpack/codec.go
@@ -117,7 +117,7 @@ func (*msgpackCodec) Unmarshal(data []byte, v any) (err error) {
 			Reason:  "UNMARSHAL_FAILED",
 			Public:  "MessagePack input exceeds size limit",
 			Private: "service/codec/msgpack.Unmarshal: len(data) exceeds maxMsgPackBytes",
-		}, errs.Int("len", int64(len(data))), errs.Int("cap", int64(maxMsgPackBytes)))
+		}, errs.Int("len", len(data)), errs.Int("cap", maxMsgPackBytes))
 	}
 	//: delegate to the library for the actual decoding.
 	uerr := gomsgpack.Unmarshal(data, v)
